+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#160](https://github.com/os2display/display-api-service/pull/160)
+  Added app:feed:list-feed-source command. Removed listing from app:feed:remove-feed-source command.
 - [#158](https://github.com/os2display/display-api-service/pull/158)
   Added thumbnails for image resources
 - [#159](https://github.com/os2display/display-api-service/pull/159)